requirements.txt
torch>=1.10.0
torchvision>=0.11.0
pandas
tqdm
Pillow
